requirements.txt
torch
torchvision
requests
opencv-python
supervision
torchaudio
transformers
tqdm
numpy
pandas
matplotlib
flask
python-dotenv
diffusers
tensorflow
scikit-learn
flask-cors

